+++ DlssUpdater/Views/Pages/GamesPage.xaml.cs
@@ -7,6 +7,7 @@ using DlssUpdater.GameLibrary;
 using DlssUpdater.Helpers;
 using DlssUpdater.Singletons;
 using DlssUpdater.ViewModels.Pages;
+using DLSSUpdater.Singletons;
 using Microsoft.Win32;
 using Wpf.Ui;
 using Wpf.Ui.Controls;
@@ -21,13 +22,15 @@ public partial class GamesPage : INavigableView<GamesViewModel>
     private readonly GameContainer _gameContainer;
     private readonly ISnackbarService _snackbar;
     private readonly DllUpdater _updater;
+    private readonly AsyncFileWatcher _fileWatcher;
 
     public GamesPage(GamesViewModel viewModel, DllUpdater updater, GameContainer gameContainer,
-        ISnackbarService snackbar)
+        ISnackbarService snackbar, AsyncFileWatcher watcher)
     {
         _updater = updater;
         _gameContainer = gameContainer;
         _snackbar = snackbar;
+        _fileWatcher = watcher;
         ViewModel = viewModel;
         DataContext = this;
 
@@ -89,6 +92,7 @@ public partial class GamesPage : INavigableView<GamesViewModel>
         if (_newGameInfo)
         {
             ViewModel.Games!.Add(ViewModel.SelectedGame!);
+            _fileWatcher.AddFile(ViewModel.SelectedGame!);
         }
         else
         {